refactor(garcom): centraliza os mapeamentos em uma única função
